+++ backend/core/cronjobs/process_queued_hooks.php
@@ -7,6 +7,9 @@
  * file that was distributed with this source code.
  */
 
+use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Filesystem\Exception\IOException;
+
 /**
  * This is the cronjob that processes the queued hooks.
  *
@@ -26,13 +29,17 @@ class BackendCoreCronjobProcessQueuedHooks extends BackendBaseCronjob
 		$db = BackendModel::getContainer()->get('database');
 
 		// create log
-		$log = new SpoonLog('custom', BACKEND_CACHE_PATH . '/logs/events');
+		$log = BackendModel::getContainer()->get('logger');
 
 		// get process-id
 		$pid = getmypid();
 
 		// store PID
-		SpoonFile::setContent(BACKEND_CACHE_PATH . '/hooks/pid', $pid);
+		$fs = new Filesystem();
+		$fs->dumpFile(
+			BACKEND_CACHE_PATH . '/hooks/pid',
+			$pid
+		);
 
 		while(true)
 		{
@@ -70,14 +77,12 @@ class BackendCoreCronjobProcessQueuedHooks extends BackendBaseCronjob
 					// reset state
 					$processedSuccessfully = false;
 
-					// logging when we are in debugmode
-					if(SPOON_DEBUG) $log->write('Callback (' . serialize($item['callback']) . ') failed.');
+					$log->info('Callback (' . serialize($item['callback']) . ') failed.');
 				}
 
 				try
 				{
-					// logging when we are in debugmode
-					if(SPOON_DEBUG) $log->write('Callback (' . serialize($item['callback']) . ') called.');
+					$log->info('Callback (' . serialize($item['callback']) . ') called.');
 
 					// call the callback
 					$return = call_user_func($item['callback'], $item['data']);
@@ -91,8 +96,7 @@ class BackendCoreCronjobProcessQueuedHooks extends BackendBaseCronjob
 						// reset state
 						$processedSuccessfully = false;
 
-						// logging when we are in debugmode
-						if(SPOON_DEBUG) $log->write('Callback (' . serialize($item['callback']) . ') failed.');
+						$log->info('Callback (' . serialize($item['callback']) . ') failed.');
 					}
 				}
 				catch(Exception $e)
@@ -117,8 +121,8 @@ class BackendCoreCronjobProcessQueuedHooks extends BackendBaseCronjob
 			// stop it
 			else
 			{
-				// remove the file
-				SpoonFile::delete(BACKEND_CACHE_PATH . '/hooks/pid');
+				$fs = new Filesystem();
+				$fs->remove(BACKEND_CACHE_PATH . '/hooks/pid');
 
 				// stop the script
 				exit;