+++ backend/core/engine/datagrid.php
@@ -17,7 +17,7 @@
  * @author Davy Hellemans <davy.hellemans@netlash.com>
  * @author Tijs Verkoyen <tijs@sumocoders.be>
  * @author Dieter Vanden Eynde <dieter@dieterve.be>
- * @author Jelmer Snoeck <jelmer.snoeck@netlash.com>
+ * @author Jelmer Snoeck <jelmer@siphoc.com>
  */
 class BackendDataGrid extends SpoonDataGrid
 {
@@ -853,6 +853,7 @@ class BackendDataGridFunctions
 	 * @param string $type The type of column. This is given since some columns can have different meanings than others.
 	 * @param string $value
 	 * @param array $attributes
+	 * @return array
 	 */
 	public static function greyOut($type, $value, array $attributes = array())
 	{