+++ backend/modules/extensions/engine/model.php
@@ -7,12 +7,16 @@
  * file that was distributed with this source code.
  */
 
+use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Filesystem\Exception\IOException;
+use Symfony\Component\Finder\Finder;
+
 /**
  * In this file we store all generic functions that we will be using in the extensions module.
  *
  * @author Dieter Vanden Eynde <dieter.vandeneynde@netlash.com>
  * @author Matthias Mullie <forkcms@mullie.eu>
- * @author Jelmer Snoeck <jelmer.snoeck@netlash.com>
+ * @author Jelmer Snoeck <jelmer@siphoc.com>
  */
 class BackendExtensionsModel
 {
@@ -215,32 +219,19 @@ class BackendExtensionsModel
 	 */
 	public static function clearCache()
 	{
-		// list of cache files to be deleted
-		$filesToDelete = array();
-
-		// backend navigation
-		$filesToDelete[] = BACKEND_CACHE_PATH . '/navigation/navigation.php';
-
-		// backend locale
-		foreach(SpoonFile::getList(BACKEND_CACHE_PATH . '/locale', '/\.php$/') as $file)
-		{
-			$filesToDelete[] = BACKEND_CACHE_PATH . '/locale/' . $file;
+		$finder = new Finder();
+		$fs = new Filesystem();
+		foreach($finder->files()
+			        ->name('*.php')
+			        ->name('*.js')
+			        ->in(BACKEND_CACHE_PATH . '/locale')
+			        ->in(FRONTEND_CACHE_PATH . '/navigation')
+			        ->in(FRONTEND_CACHE_PATH . '/locale')
+		        as $file
+		) {
+			$fs->remove($file->getRealPath());
 		}
-
-		// frontend navigation
-		foreach(SpoonFile::getList(FRONTEND_CACHE_PATH . '/navigation', '/\.(php|js)$/') as $file)
-		{
-			$filesToDelete[] = FRONTEND_CACHE_PATH . '/navigation/' . $file;
-		}
-
-		// frontend locale
-		foreach(SpoonFile::getList(FRONTEND_CACHE_PATH . '/locale', '/\.php$/') as $file)
-		{
-			$filesToDelete[] = FRONTEND_CACHE_PATH . '/locale/' . $file;
-		}
-
-		// delete the files
-		foreach($filesToDelete as $file) SpoonFile::delete($file);
+		$fs->remove(BACKEND_CACHE_PATH . '/navigation/navigation.php');
 	}
 
 	/**
@@ -301,9 +292,7 @@ class BackendExtensionsModel
 	public static function existsModule($module)
 	{
 		$module = (string) $module;
-
-		// check if modules directory exists
-		return SpoonDirectory::exists(BACKEND_MODULES_PATH . '/' . $module);
+		return is_dir(BACKEND_MODULES_PATH . '/' . $module);
 	}
 
 	/**
@@ -333,9 +322,7 @@ class BackendExtensionsModel
 	public static function existsTheme($theme)
 	{
 		$theme = (string) $theme;
-
-		// check if modules directory exists
-		return SpoonDirectory::exists(FRONTEND_PATH . '/themes/' . $theme) || $theme == 'core';
+		return is_dir(FRONTEND_PATH . '/themes/' . $theme) || $theme == 'core';
 	}
 
 	/**
@@ -466,7 +453,7 @@ class BackendExtensionsModel
 		$information = array('data' => array(), 'warnings' => array());
 
 		// information needs to exists
-		if(SpoonFile::exists($pathInfoXml))
+		if(is_file($pathInfoXml))
 		{
 			try
 			{
@@ -533,7 +520,7 @@ class BackendExtensionsModel
 		$installedModules = (array) BackendModel::getContainer()->get('database')->getRecords('SELECT name FROM modules', null, 'name');
 
 		// get modules present on the filesystem
-		$modules = SpoonDirectory::getList(BACKEND_MODULES_PATH, false, null, '/^[a-zA-Z0-9_]+$/');
+		$modules = BackendModel::getModulesOnFilesystem(false);
 
 		// all modules that are managable in the backend
 		$manageableModules = array();
@@ -737,21 +724,22 @@ class BackendExtensionsModel
 	 */
 	public static function getThemes()
 	{
-		// fetch themes
-		$records = (array) SpoonDirectory::getList(FRONTEND_PATH . '/themes/', false, array('.svn'));
+		$records = array();
+		$records['core'] = array(
+			'value' => 'core',
+			'label' => BL::lbl('NoTheme'),
+			'thumbnail' => '/frontend/core/layout/images/thumbnail.png',
+			'installed' => self::isThemeInstalled('core'),
+			'installable' => false,
+		);
 
-		// loop and complete the records
-		foreach($records as $key => $record)
+		$finder = new Finder();
+		foreach($finder->directories()->in(FRONTEND_PATH . '/themes')->depth(0) as $directory)
 		{
 			try
 			{
-				// path to info.xml
-				$pathInfoXml = PATH_WWW . '/frontend/themes/' . $record . '/info.xml';
-
-				// load info.xml
+				$pathInfoXml = PATH_WWW . '/frontend/themes/' . $directory->getBasename() . '/info.xml';
 				$infoXml = @new SimpleXMLElement($pathInfoXml, LIBXML_NOCDATA, true);
-
-				// convert xml to useful array
 				$information = self::processThemeXml($infoXml);
 				if(!$information) throw new BackendException('Invalid info.xml');
 			}
@@ -763,27 +751,15 @@ class BackendExtensionsModel
 				$information['thumbnail'] = 'thumbnail.png';
 			}
 
-			// add additional values
-			$records[$record]['value'] = $record;
-			$records[$record]['label'] = $record;
-			$records[$record]['thumbnail'] = '/frontend/themes/' . $record . '/' . $information['thumbnail'];
+			$item = array();
+			$item['value'] = $directory->getBasename();
+			$item['label'] = $directory->getBasename();
+			$item['thumbnail'] =  '/frontend/themes/' . $item['value'] . '/' . $information['thumbnail'];
+			$item['installed'] = self::isThemeInstalled($item['value']);
+			$item['installable'] = isset($information['templates']);
 
-			// doublecheck if templates for this theme are installed already
-			$records[$record]['installed'] = self::isThemeInstalled($record);
-			$records[$record]['installable'] = isset($information['templates']);
-
-			// unset the key
-			unset($records[$key]);
+			$records[$item['value']] = $item;
 		}
-
-		// add core theme
-		$core = array('core' => array());
-		$core['core']['value'] = 'core';
-		$core['core']['label'] = BL::lbl('NoTheme');
-		$core['core']['thumbnail'] = '/frontend/core/layout/images/thumbnail.png';
-		$core['core']['installed'] = self::isThemeInstalled('core');
-		$core['core']['installable'] = false;
-		$records = array_merge($core, $records);
 
 		return (array) $records;
 	}
@@ -999,7 +975,7 @@ class BackendExtensionsModel
 		if($return === false) return false;
 
 		// unlink the random file
-		SpoonFile::delete($path . '/' . $file);
+		unlink($path . '/' . $file);
 
 		return true;
 	}