+++ frontend/core/engine/ajax.php
@@ -7,6 +7,8 @@
  * file that was distributed with this source code.
  */
 
+use Symfony\Component\Finder\Finder;
+
 /**
  * FrontendAJAX
  * This class will handle AJAX-related stuff
@@ -121,24 +123,20 @@ class FrontendAJAX extends KernelLoader implements ApplicationInterface
 	{
 		// check if module is set
 		if($this->getModule() === null) throw new BackendException('Module has not yet been set.');
-	
-		// build the path (core is a special case)
-		if($this->getModule() == 'core') $path = FRONTEND_PATH . '/core/ajax/';
-		else $path = FRONTEND_PATH . '/modules/' . $this->getModule() . '/ajax/';
-	
-		// get the possible actions
-		$possibleActions = SpoonFile::getList($path);
-		
-		// validate
-		if(!in_array($value.'.php', $possibleActions))
-		{
-			// create fake action
-			$fakeAction = new FrontendBaseAJAXAction('', '');
 
-			// output error
+		// grab the file
+		$finder = new Finder();
+		$finder->name($value . '.php');
+		if($this->getModule() == 'core') $finder->in(FRONTEND_PATH . '/core/ajax/');
+		else $finder->in(FRONTEND_PATH . '/modules/' . $this->getModule() . '/ajax/');
+
+		// validate
+		if(count($finder->files()) != 1)
+		{
+			$fakeAction = new FrontendBaseAJAXAction('', '');
 			$fakeAction->output(FrontendBaseAJAXAction::BAD_REQUEST, null, 'Action not correct.');
 		}
-	
+
 		// set property
 		$this->action = (string) $value;
 	}
@@ -185,7 +183,7 @@ class FrontendAJAX extends KernelLoader implements ApplicationInterface
 	{
 		// get the possible modules
 		$possibleModules = FrontendModel::getModules();
-		
+
 		// validate
 		if(!in_array($value, $possibleModules))
 		{
@@ -195,7 +193,7 @@ class FrontendAJAX extends KernelLoader implements ApplicationInterface
 			// output error
 			$fakeAction->output(FrontendBaseAJAXAction::BAD_REQUEST, null, 'Module not correct.');
 		}
-	
+
 		// set property
 		$this->module = (string) $value;
 	}
@@ -259,7 +257,7 @@ class FrontendAJAXAction extends FrontendBaseAJAXAction
 		else $path = FRONTEND_PATH . '/modules/' . $this->getModule() . '/ajax/' . $this->getAction() . '.php';
 
 		// check if the config is present? If it isn't present there is a huge problem, so we will stop our code by throwing an error
-		if(!SpoonFile::exists($path)) throw new FrontendException('The actionfile (' . $path . ') can\'t be found.');
+		if(!is_file($path)) throw new FrontendException('The actionfile (' . $path . ') can\'t be found.');
 
 		// require the ajax file, we know it is there because we validated it before (possible actions are defined by existance of the file).
 		require_once $path;
@@ -315,7 +313,9 @@ class FrontendAJAXAction extends FrontendBaseAJAXAction
 		else $frontendModulePath = FRONTEND_MODULES_PATH . '/' . $this->getModule();
 
 		// check if the config is present? If it isn't present there is a huge problem, so we will stop our code by throwing an error
-		if(!SpoonFile::exists($frontendModulePath . '/config.php')) throw new FrontendException('The configfile for the module (' . $this->getModule() . ') can\'t be found.');
+		if(!is_file($frontendModulePath . '/config.php')) {
+			throw new FrontendException('The configfile for the module (' . $this->getModule() . ') can\'t be found.');
+		}
 
 		// build config-object-name
 		$configClassName = 'Frontend' . SpoonFilter::toCamelCase($this->getModule() . '_config');
@@ -324,7 +324,9 @@ class FrontendAJAXAction extends FrontendBaseAJAXAction
 		require_once $frontendModulePath . '/config.php';
 
 		// validate if class exists (aka has correct name)
-		if(!class_exists($configClassName)) throw new FrontendException('The config file is present, but the classname should be: ' . $configClassName . '.');
+		if(!class_exists($configClassName)) {
+			throw new FrontendException('The config file is present, but the classname should be: ' . $configClassName . '.');
+		}
 
 		// create config-object, the constructor will do some magic
 		$this->config = new $configClassName($this->getModule());