+++ frontend/core/engine/model.php
@@ -1,13 +1,16 @@
 <?php
 
-use \TijsVerkoyen\Akismet\Akismet;
-
 /*
  * This file is part of Fork CMS.
  *
  * For the full copyright and license information, please view the license
  * file that was distributed with this source code.
  */
+
+use TijsVerkoyen\Akismet\Akismet;
+use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Filesystem\Exception\IOException;
+use Symfony\Component\Finder\Finder;
 
 require_once __DIR__ . '/../../../app/BaseModel.php';
 
@@ -225,17 +228,17 @@ class FrontendModel extends BaseModel
 	 * @param string $path The path wherein the thumbnail-folders will be stored.
 	 * @param string $sourceFile The location of the source file.
 	 */
-	public static function generateThumbnails($path, $sourcefile)
+	public static function generateThumbnails($path, $sourceFile)
 	{
 		// get folder listing
 		$folders = self::getThumbnailFolders($path);
-		$filename = basename($sourcefile);
+		$filename = basename($sourceFile);
 
 		// loop folders
 		foreach($folders as $folder)
 		{
 			// generate the thumbnail
-			$thumbnail = new SpoonThumbnail($sourcefile, $folder['width'], $folder['height']);
+			$thumbnail = new SpoonThumbnail($sourceFile, $folder['width'], $folder['height']);
 			$thumbnail->setAllowEnlargement(true);
 
 			// if the width & height are specified we should ignore the aspect ratio
@@ -474,30 +477,26 @@ class FrontendModel extends BaseModel
 	 */
 	public static function getThumbnailFolders($path, $includeSource = false)
 	{
-		$folders = SpoonDirectory::getList((string) $path, false, null, '/^([0-9]*)x([0-9]*)$/');
-
-		if($includeSource && SpoonDirectory::exists($path . '/source')) $folders[] = 'source';
-
 		$return = array();
+		$fs = new Filesystem();
+		if(!$fs->exists($path)) return $return;
+		$finder = new Finder();
+		$finder->name('/^([0-9]*)x([0-9]*)$/');
+		if($includeSource) $finder->name('source');
 
-		foreach($folders as $folder)
-		{
-			$item = array();
-			$chunks = explode('x', $folder, 2);
-
-			// skip invalid items
+		foreach($finder->directories()->in($path) as $directory) {
+			$chunks = explode('x', $directory->getBasename(), 2);
 			if(count($chunks) != 2 && !$includeSource) continue;
 
-			$item['dirname'] = $folder;
-			$item['path'] = $path . '/' . $folder;
-			if(substr($path, 0, strlen(PATH_WWW)) == PATH_WWW) $item['url'] = substr($item['path'], strlen(PATH_WWW));
-			if($folder == 'source')
-			{
+			$item = array();
+			$item['dirname'] = $directory->getBasename();
+			$item['path'] = $directory->getRealPath();
+			if(substr($path, 0, strlen(PATH_WWW)) == PATH_WWW) $item['url'] = substr($path, strlen(PATH_WWW));
+
+			if($item['dirname'] == 'source') {
 				$item['width'] = null;
 				$item['height'] = null;
-			}
-			else
-			{
+			} else {
 				$item['width'] = ($chunks[0] != '') ? (int) $chunks[0] : null;
 				$item['height'] = ($chunks[1] != '') ? (int) $chunks[1] : null;
 			}
@@ -759,11 +758,12 @@ class FrontendModel extends BaseModel
 	 */
 	public static function startProcessingHooks()
 	{
+		$fs = new Filesystem();
 		// is the queue already running?
-		if(SpoonFile::exists(FRONTEND_CACHE_PATH . '/hooks/pid'))
+		if($fs->exists(FRONTEND_CACHE_PATH . '/hooks/pid'))
 		{
 			// get the pid
-			$pid = trim(SpoonFile::getContent(FRONTEND_CACHE_PATH . '/hooks/pid'));
+			$pid = trim(file_get_contents(FRONTEND_CACHE_PATH . '/hooks/pid'));
 
 			// running on windows?
 			if(strtolower(substr(php_uname('s'), 0, 3)) == 'win')
@@ -775,7 +775,7 @@ class FrontendModel extends BaseModel
 				if($output == '' || $output === false)
 				{
 					// delete the pid file
-					SpoonFile::delete(FRONTEND_CACHE_PATH . '/hooks/pid');
+					$fs->remove(FRONTEND_CACHE_PATH . '/hooks/pid');
 				}
 
 				// already running
@@ -792,7 +792,7 @@ class FrontendModel extends BaseModel
 				if($output === false)
 				{
 					// delete the pid file
-					SpoonFile::delete(FRONTEND_CACHE_PATH . '/hooks/pid');
+					$fs->remove(FRONTEND_CACHE_PATH . '/hooks/pid');
 				}
 
 				// already running
@@ -803,10 +803,10 @@ class FrontendModel extends BaseModel
 			else
 			{
 				// check if the process is still running, by checking the proc folder
-				if(!SpoonFile::exists('/proc/' . $pid))
+				if(!$fs->exists('/proc/' . $pid))
 				{
 					// delete the pid file
-					SpoonFile::delete(FRONTEND_CACHE_PATH . '/hooks/pid');
+					$fs->remove(FRONTEND_CACHE_PATH . '/hooks/pid');
 				}
 
 				// already running
@@ -892,10 +892,8 @@ class FrontendModel extends BaseModel
 		$eventName = (string) $eventName;
 
 		// create log instance
-		$log = new SpoonLog('custom', PATH_WWW . '/frontend/cache/logs/events');
-
-		// logging when we are in debugmode
-		if(SPOON_DEBUG) $log->write('Event (' . $module . '/' . $eventName . ') triggered.');
+		$log = self::getContainer()->get('logger');
+		$log->info('Event (' . $module . '/' . $eventName . ') triggered.');
 
 		// get all items that subscribe to this event
 		$subscriptions = (array) self::getContainer()->get('database')->getRecords(
@@ -924,8 +922,7 @@ class FrontendModel extends BaseModel
 				// add
 				$queuedItems[] = self::getContainer()->get('database')->insert('hooks_queue', $item);
 
-				// logging when we are in debugmode
-				if(SPOON_DEBUG) $log->write('Callback (' . $subscription['callback'] . ') is subscribed to event (' . $module . '/' . $eventName . ').');
+				$log->info('Callback (' . $subscription['callback'] . ') is subscribed to event (' . $module . '/' . $eventName . ').');
 			}
 
 			// start processing