+++ frontend/modules/faq/actions/category.php
@@ -11,7 +11,7 @@
  * This is the category-action
  *
  * @author Lester Lievens <lester.lievens@netlash.com>
- * @author Jelmer Snoeck <jelmer.snoeck@netlash.com>
+ * @author Jelmer Snoeck <jelmer@siphoc.com>
  */
 class FrontendFaqCategory extends FrontendBaseBlock
 {