+++ frontend/modules/location/actions/index.php
@@ -11,7 +11,7 @@
  * This is the index-action, it has an overview of locations.
  *
  * @author Matthias Mullie <forkcms@mullie.eu>
- * @author Jelmer Snoeck <jelmer.snoeck@netlash.com>
+ * @author Jelmer Snoeck <jelmer@siphoc.com>
  * @author Tijs Verkoyen <tijs@sumocoders.be>
  */
 class FrontendLocationIndex extends FrontendBaseBlock