+++ frontend/modules/location/widgets/location.php
@@ -11,7 +11,7 @@
  * This is the location-widget: 1 specific address
  *
  * @author Matthias Mullie <forkcms@mullie.eu>
- * @author Jelmer Snoeck <jelmer.snoeck@netlash.com>
+ * @author Jelmer Snoeck <jelmer@siphoc.com>
  * @author Tijs Verkoyen <tijs@sumocoders.be>
  */
 class FrontendLocationWidgetLocation extends FrontendBaseWidget