+++ frontend/modules/search/actions/index.php
@@ -7,6 +7,10 @@
  * file that was distributed with this source code.
  */
 
+use Symfony\Component\Finder\Finder;
+use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Filesystem\Exception\IOException;
+
 /**
  * This action will display a form to search
  *
@@ -120,7 +124,7 @@ class FrontendSearchIndex extends FrontendBaseBlock
 		if(SPOON_DEBUG) return false;
 
 		// check if cachefile exists
-		if(!SpoonFile::exists($this->cacheFile)) return false;
+		if(!is_file($this->cacheFile)) return false;
 
 		// get cachefile modification time
 		$cacheInfo = @filemtime($this->cacheFile);
@@ -172,7 +176,11 @@ class FrontendSearchIndex extends FrontendBaseBlock
 		if(!SPOON_DEBUG)
 		{
 			// set cache content
-			SpoonFile::setContent($this->cacheFile, "<?php\n" . '$pagination = ' . var_export($this->pagination, true) . ";\n" . '$items = ' . var_export($this->items, true) . ";\n?>");
+			$fs = new Filesystem();
+			$fs->dumpFile(
+				$this->cacheFile,
+				"<?php\n" . '$pagination = ' . var_export($this->pagination, true) . ";\n" . '$items = ' . var_export($this->items, true) . ";\n?>"
+			);
 		}
 	}